refactor: 기존 스키마 schema 폴더로 이동
