+++ src/pages/@owner/menu/hooks/use-form-validation.ts
@@ -6,15 +6,14 @@ import {
   MENU_DESCRIPTION_VALIDATOR,
   MENU_PRICE_VALIDATOR,
   MENU_IMAGE_VALIDATOR,
-} from '@pages/@owner/menu/utils/menu-form-validator.schema';
+} from '@pages/@owner/menu/schema/menu-form-validator.schema';
 
-const menuSchema = z
-  .object({
-    name: MENU_NAME_VALIDATOR,
-    description: MENU_DESCRIPTION_VALIDATOR,
-    price: MENU_PRICE_VALIDATOR,
-    imageUrl: MENU_IMAGE_VALIDATOR,
-  });
+const menuSchema = z.object({
+  name: MENU_NAME_VALIDATOR,
+  description: MENU_DESCRIPTION_VALIDATOR,
+  price: MENU_PRICE_VALIDATOR,
+  imageUrl: MENU_IMAGE_VALIDATOR,
+});
 
 export type MenuFormData = z.infer<typeof menuSchema>;
 
@@ -41,7 +40,10 @@ export const useFormValidation = (initialData?: Partial<MenuFormData>) => {
   };
 
   const updateDescription = (description: string) => {
-    setValue('description', description, { shouldValidate: true, shouldDirty: true });
+    setValue('description', description, {
+      shouldValidate: true,
+      shouldDirty: true,
+    });
   };
 
   const updatePrice = (price: string) => {
@@ -51,11 +53,17 @@ export const useFormValidation = (initialData?: Partial<MenuFormData>) => {
       return;
     }
     const numericPrice = Number(numbersOnly);
-    setValue('price', numericPrice, { shouldValidate: true, shouldDirty: true });
+    setValue('price', numericPrice, {
+      shouldValidate: true,
+      shouldDirty: true,
+    });
   };
 
   const updateImageUrl = (image: File | null) => {
-    setValue('imageUrl', image || undefined, { shouldValidate: true, shouldDirty: true });
+    setValue('imageUrl', image || undefined, {
+      shouldValidate: true,
+      shouldDirty: true,
+    });
   };
 
   const Errors = {
